+++ my-app/src/components/Warning.jsx
@@ -1,10 +1,12 @@
 import styled from "styled-components";
 
-export const WarningItem = styled.strong`
-    display: block;
+const Warning = styled.strong`
+    display: none;
     margin-top: 6px;
     font-size: 12px;
     font-weight: 400;
     line-height: 14px;
     color: #eb5757;
-`;
+`
+
+export default Warning;
