--- my-app/src/components/shared/Warning/Warning.jsx
@@ -1,8 +1,0 @@
-import React from "react";
-import { WarningItem } from "./warning.style";
-
-export default function Warning({ message, visible }) {
-    if (visible) {
-        return <WarningItem>*{message}</WarningItem>;
-    }
-}
